Combine cyclic import into single file
