+++ interpreter.cabal
@@ -1,6 +1,6 @@
 cabal-version: 2.2
 
--- This file has been generated from package.yaml by hpack version 0.38.0.
+-- This file has been generated from package.yaml by hpack version 0.37.0.
 --
 -- see: https://github.com/sol/hpack
 
@@ -28,9 +28,7 @@ library
       Interpreter
       Interpreter.Arithmetic
       Interpreter.Comparison
-      Interpreter.Control
       Interpreter.Core
-      Interpreter.HigherOrder
       Interpreter.IO
       Interpreter.List
       Interpreter.Stack
