--- interpreter.cabal.old
@@ -1,86 +1,0 @@
-cabal-version: 2.2
-
--- This file has been generated from package.yaml by hpack version 0.38.0.
---
--- see: https://github.com/sol/hpack
-
-name:           interpreter
-version:        0.1.0.0
-description:    Please see the README on GitHub at <https://github.com/githubuser/interpreter#readme>
-homepage:       https://github.com/githubuser/interpreter#readme
-bug-reports:    https://github.com/githubuser/interpreter/issues
-author:         Author name here
-maintainer:     example@example.com
-copyright:      2025 Author name here
-license:        BSD-3-Clause
-license-file:   LICENSE
-build-type:     Simple
-extra-source-files:
-    README.md
-    CHANGELOG.md
-
-source-repository head
-  type: git
-  location: https://github.com/githubuser/interpreter
-
-library
-  exposed-modules:
-      Interpreter
-      Lib
-      Parser
-      Types
-      Tokenizer
-  other-modules:
-      Paths_interpreter
-  autogen-modules:
-      Paths_interpreter
-  hs-source-dirs:
-      src
-  ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-export-lists -Wmissing-home-modules -Wpartial-fields -Wredundant-constraints
-  build-depends:
-      base >=4.7 && <5
-    , bytestring
-    , containers
-    , directory
-    , extra
-    , mtl
-  default-language: Haskell2010
-
-executable interpreter-exe
-  main-is: Main.hs
-  other-modules:
-      Paths_interpreter
-  autogen-modules:
-      Paths_interpreter
-  hs-source-dirs:
-      app
-  ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-export-lists -Wmissing-home-modules -Wpartial-fields -Wredundant-constraints -threaded -rtsopts -with-rtsopts=-N
-  build-depends:
-      base >=4.7 && <5
-    , bytestring
-    , containers
-    , directory
-    , extra
-    , interpreter
-    , mtl
-  default-language: Haskell2010
-
-test-suite interpreter-test
-  type: exitcode-stdio-1.0
-  main-is: Spec.hs
-  other-modules:
-      Paths_interpreter
-  autogen-modules:
-      Paths_interpreter
-  hs-source-dirs:
-      test
-  ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-export-lists -Wmissing-home-modules -Wpartial-fields -Wredundant-constraints -threaded -rtsopts -with-rtsopts=-N
-  build-depends:
-      base >=4.7 && <5
-    , bytestring
-    , containers
-    , directory
-    , extra
-    , interpreter
-    , mtl
-  default-language: Haskell2010
