+++ src/Interpreter/TokenExecution.hs
@@ -5,11 +5,48 @@ module Interpreter.TokenExecution
         executeToken,
         executeValue,
         executeQuotationOrValue,
-        executeExec
+        executeExec,
+
+        -- Control flow
+        executeIf,
+        executeTimes,
+        executeLoop,
+        executeLoopWithComponents,
+        execTimesIterative,
+
+        -- Higher order functions
+        applyEach,
+        executeMap,
+        executeEach,
+        executeFoldl,
+        applyToItem,
+        applyFunction,
+        applyToElement,
+        mapListWithQuotation,
+        foldListWithQuotation,
+
+        -- Higher order functions
+        applyEach,
+        executeMap,
+        executeEach,
+        executeFoldl,
+        applyToItem,
+        applyFunction,
+        applyToElement,
+        mapListWithQuotation,
+        foldListWithQuotation
     ) where
 
 import Types 
-import Interpreter.Control
+import Interpreter.Stack
+import Data.Map as Map
+import Interpreter.Variables
+import Interpreter.String
+import Interpreter.List
+import Interpreter.IO
+import Interpreter.Comparison
+import Interpreter.Arithmetic
+import Control.Monad
 
 -- | Execute a stream of tokens
 executeTokenStream :: [Token] -> State -> Either ProgramError (State, [Token])
@@ -82,3 +119,235 @@ executeExec state = do
             (finalState, _) <- executeTokenStream tokens state'
             return finalState
         _ -> Left $ ExpectedQuotation quotation
+
+-- | Execute if operation - handles the token stream
+executeIf :: [Token] -> State -> Either ProgramError (State, [Token])
+executeIf tokens state = do
+    (condition, state') <- popValue state
+    (thenBranch, afterThen) <- splitAtQuotation tokens
+    (elseBranch, rest) <- splitAtQuotation afterThen
+    
+    case condition of
+        BoolValue True -> do
+            (finalState, _) <- executeTokenStream thenBranch state'
+            return (finalState, rest)
+        BoolValue False -> do
+            (finalState, _) <- executeTokenStream elseBranch state'
+            return (finalState, rest)
+        _ -> Left $ ExpectedBool condition
+
+-- | Execute times operation - handles the token stream
+executeTimes :: [Token] -> State -> Either ProgramError (State, [Token])
+executeTimes tokens state = do
+    -- Pop the count and get the code block
+    (countValue, state') <- popValue state
+    (block, rest) <- splitAtQuotation tokens
+    
+    case countValue of
+        IntValue count ->
+            if count <= 0
+                then Right (state', rest)
+                else do
+                    finalState <- execTimesIterative block count state'
+                    return (finalState, rest)
+        _ -> Left $ ExpectedBoolOrNumber countValue
+
+-- | Execute loop operation - handles the token stream
+executeLoop :: [Token] -> State -> Either ProgramError (State, [Token])
+executeLoop tokens state = do
+    -- Pop the initial value and get break condition and body from token stream
+    (initial, state') <- popValue state
+    (condTokens, afterCond) <- splitAtQuotation tokens
+    (bodyTokens, rest) <- splitAtQuotation afterCond
+    
+    -- Execute the loop with these components
+    finalState <- executeLoopWithComponents initial condTokens bodyTokens state'
+    return (finalState, rest)
+
+-- | Execute a loop with specified components
+executeLoopWithComponents :: Value -> [Token] -> [Token] -> State -> Either ProgramError State
+executeLoopWithComponents initial condTokens bodyTokens initState =
+    let initialState' = pushValue initial initState
+        loop currentState = do
+            -- Execute the condition and check top of stack
+            (condState, _) <- executeTokenStream condTokens currentState
+            case stack condState of
+                (BoolValue True:restStack) ->
+                    -- Condition is true, exit the loop
+                    Right $ condState { stack = restStack }
+                (BoolValue False:restStack) -> do
+                    -- Condition is false, execute body and continue
+                    let stateAfterCond = condState { stack = restStack }
+                    (bodyState, _) <- executeTokenStream bodyTokens stateAfterCond
+                    loop bodyState
+                (invalidValue:_) ->
+                    Left (ExpectedBool invalidValue)
+                [] ->
+                    Left StackEmpty
+    in loop initialState'
+
+-- | Iterative implementation of times operation
+execTimesIterative :: [Token] -> Integer -> State -> Either ProgramError State
+execTimesIterative tokens count state =
+    let loop 0 s = Right s
+        loop n s = do
+            (s', _) <- executeTokenStream tokens s
+            loop (n-1) s'
+    in loop count state
+
+-- | Apply each operation to a single item
+applyEach :: Value -> [Token] -> State -> Either ProgramError State
+applyEach item tokens state = do
+    let state' = pushValue item state
+    (resultState, _) <- executeTokenStream tokens state'
+    return resultState
+
+-- | Execute each operation
+executeEach :: [Token] -> State -> Either ProgramError (State, [Token])
+--executeEach [] state = Right (state, [])
+executeEach (token:tokens) state = do
+    (quotationTokens, remainingTokens) <- splitAtQuotation tokens
+    (listValue, state') <- popValue state
+    case listValue of
+        ListValue values -> do
+            finalState <- foldM (applyToElement quotationTokens) state' values
+            return (finalState, remainingTokens)
+        _ -> Left $ ExpectedList listValue
+
+-- | Execute map operation
+executeMap :: [Token] -> State -> Either ProgramError (State, [Token])
+executeMap tokens state = do
+    (quotationTokens, remainingTokens) <- splitAtQuotation tokens
+    (listValue, state') <- popValue state
+    case listValue of
+        ListValue values -> do
+            -- Map over function, with function gotten from tokens
+            result <- mapM (applyFunction quotationTokens) values
+            let finalState = pushValue (ListValue result) state'
+            return (finalState, remainingTokens)
+        _ -> Left $ ExpectedList listValue
+
+-- | Execute foldl operation
+executeFoldl :: State -> Either ProgramError State
+executeFoldl state = do
+    (quotation, state1) <- popValue state
+    (initial, state2) <- popValue state1
+    (list, state3) <- popValue state2
+    case list of
+        ListValue values ->
+            case quotation of
+                QuotationValue tokens -> 
+                    foldListWithQuotation values initial tokens state3
+                _ -> 
+                    foldListWithQuotation values initial [ValueToken quotation] state3
+        _ -> Left $ ExpectedList list
+
+-- | Apply function to a single item
+applyToItem :: [Token] -> State -> Value -> Either ProgramError Value
+applyToItem tokens state item = do
+    -- Push item onto a clean state (starting with original state but empty stack)
+    let itemState = pushValue item state { stack = [] }
+    -- Execute the tokens on this state
+    (resultState, _) <- executeTokenStream tokens itemState
+    -- Get the result from the top of the stack
+    case stack resultState of
+        (result:_) -> Right result
+        [] -> Left StackEmpty
+
+-- | Apply a function to a single value
+applyFunction :: [Token] -> Value -> Either ProgramError Value
+applyFunction tokens value = do
+    -- Duplicate state to make a copy of the stack
+    let tempState = State { dictionary = Map.empty, stack = [value]}
+    (resultState, _) <- executeTokenStream tokens tempState
+    case stack resultState of
+        (result:_) -> Right result
+        [] -> Left StackEmpty
+
+-- | Apply a function to a single element and update the state
+applyToElement :: [Token] -> State -> Value -> Either ProgramError State
+applyToElement tokens state element = do
+    let stateWithElement = pushValue element state
+    -- Execute the tokens in new state
+    (resultState, _) <- executeTokenStream tokens stateWithElement
+    return resultState
+
+-- | Map a list with a quotation
+mapListWithQuotation :: [Value] -> [Token] -> State -> Either ProgramError State
+mapListWithQuotation values tokens state = do
+    -- Process each value in the list, collecting results
+    results <- mapM (applyToItem tokens state) values
+    -- Push the list of results onto the stack
+    return $ pushValue (ListValue results) state
+
+-- | Fold a list with a quotation
+foldListWithQuotation :: [Value] -> Value -> [Token] -> State -> Either ProgramError State
+foldListWithQuotation values initial tokens state =
+    let foldItem acc item = do
+            (accValue, accState) <- acc
+            let state' = pushValue item $ pushValue accValue accState
+            (resultState, _) <- executeTokenStream tokens state'
+            case stack resultState of
+                (result:rest) -> Right (result, resultState { stack = rest })
+                [] -> Left StackEmpty
+    in foldM (\(acc, s) item -> foldItem (Right (acc, s)) item) (initial, state) values >>= \(result, finalState) ->
+        return $ pushValue result finalState
+
+-- | Execute the program (list of tokens) with given state
+executeProgram :: [Token] -> State -> Either ProgramError (State, Value)
+executeProgram tokens state = do
+    (finalState, _) <- executeTokenStream tokens state
+    (finalState', value) <- extractFinalValue finalState
+    let evaluatedValue = evaluateValue value finalState'
+    return (finalState, evaluatedValue)
+
+-- | Execute operator token
+executeOperator :: String -> State -> Either ProgramError State
+executeOperator op = case op of
+    "+" -> executeArithmetic (+)
+    "-" -> executeArithmetic (-)
+    "*" -> executeArithmetic (*)
+    "div" -> executeIntegerDivision
+    "/" -> executeFloatDivision
+    "<" -> executeComparison (<)
+    ">" -> executeComparison (>)
+    "==" -> executeEquality
+    "&&" -> executeLogical (&&)
+    "||" -> executeLogical (||)
+    "not" -> executeNot
+    "dup" -> executeDup
+    "swap" -> executeSwap
+    "pop" -> executePop
+    "parseInteger" -> executeParseInteger
+    "parseFloat" -> executeParseFloat
+    "words" -> executeWords
+    "head" -> executeHead
+    "tail" -> executeTail
+    "empty" -> executeEmpty
+    "length" -> executeLength
+    "cons" -> executeCons
+    "append" -> executeAppend
+    "print" -> executePrint
+    "read" -> executeRead
+    _ -> \s -> Left $ UnknownSymbol op
+
+-- | Extract the final value from state
+extractFinalValue :: State -> Either ProgramError (State, Value)
+extractFinalValue state = case stack state of
+    [] -> Left ProgramFinishedWithNoValues
+    [v] -> Right (state, v)
+    vs -> Left (ProgramFinishedWithMultipleValues vs)
+
+-- | Evaluate a value, resolving symbols in lists
+evaluateValue :: Value -> State -> Value
+evaluateValue (ListValue lv) state = ListValue (evaluateList lv state)
+evaluateValue value _ = value
+
+-- | Evaluate a list by replacing symbols with their values
+evaluateList :: [Value] -> State -> [Value]
+evaluateList [] _ = []
+evaluateList (SymbolValue name : rest) state =
+    case Map.lookup name (dictionary state) of
+        Just value -> value : evaluateList rest state
+        Nothing -> SymbolValue name : evaluateList rest state
+evaluateList (value : rest) state = value : evaluateList rest state
